complete file refactor for simchat
