+++ crates/tools/sim_chat/src/lib3h_simchat/handle_lib3h_event.rs
@@ -1,5 +1,5 @@
 
-use crate::current_timestamp;
+use super::current_timestamp;
 use crate::simchat::{ChatEvent, SimChatMessage};
 use lib3h_protocol::{
     data_types::{QueryEntryData, QueryEntryResultData},
