+++ crates/tools/sim_chat/src/lib3h_simchat/tests/mock_engine.rs
@@ -1,4 +1,4 @@
-use detach::Detach;
+use detach::{Detach, detach_run};
 use lib3h::{engine::ghost_engine::ClientToLib3hMessage, error::Lib3hError};
 use lib3h_ghost_actor::{
     create_ghost_channel, prelude::WorkWasDone, GhostActor, GhostCanTrack, GhostContextEndpoint,
